[GSCOLLECT-696] Set up IntelliJ Inspections and fix violations.
git-svn-id: svn+ssh://gscollections.svn.services.gs.com/svnroot/gscollections-svn/trunk@72 d5c9223b-1aff-41ac-aadd-f810b4a99ac4

+++ collections/src/main/java/com/gs/collections/impl/lazy/primitive/CollectFloatIterable.java
@@ -152,7 +152,7 @@ public class CollectFloatIterable<T>
 
     public double average()
     {
-        return (double) this.sum() / (double) this.size();
+        return this.sum() / (double) this.size();
     }
 
     public double median()
